% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_binning.R
\name{bin_init}
\alias{bin_init}
\title{Initial binning for all variables}
\usage{
bin_init(sdf, x, y, init_bins = 100)
}
\arguments{
\item{sdf}{Spark data frame}

\item{x}{the variable to bin}

\item{y}{the target variable, which should take only values 0 and 1,
no missing should be allowed}

\item{init_bins}{number of initial bins for numeric variables
for character variables, init_bins is equal to 1/minp
where minp is the minimum frequency of a level, if the frequency is less than
minp then the corresponding level will be grouped to '_Other_'}
}
\value{
for numeric variable x, it returns a numeric binning object with class name "intervalbin" with the
following fields:
cuts - the break points (left end should be -Inf, right end should be Inf)
good - number of good (0) in each interval
bad - number of bad (1) in each interval
missing - a list with two components for missing values: good count and
bad count
for character variable x, it returns a character binning object with class name "nominalbin" with the
following fields:
xlevels - including major levels (> minp) and two special levels
"_Missing_" and "_Other_"
ylevels - mapped new levels, in the initial binning it should be same as
xlevels
good - number of good in each xlevel
bad - number of bad in each xlevel

EXAMPLE of returned object
obj <- list(xlevels = c("A", "B", "_Missing_", "_Other_"),
            ylevels = c("A", "B", "_Missing_", "_Other_"),
            good = c(200, 200, 200, 200),
            bad = c(100, 200, 300, 400),
            minp = 0.01)
}
\description{
Initial binning for all variables
}
