% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_binning.R
\name{bin_init_char}
\alias{bin_init_char}
\title{Initial binning for character variables}
\usage{
bin_init_char(sdf, x, y, minp = 0.01)
}
\arguments{
\item{sdf}{Spark data frame}

\item{x}{the character variable to bin}

\item{y}{the target variable, which should take only values 0 and 1,
no missing should be allowed}

\item{minp}{the minimum frequency of a level, if the frequency is less than
minp then the corresponding level will be grouped to '_Other_'}
}
\value{
a character binning object with class name "nominalbin" with the
following fields:
xlevels - including major levels (> minp) and two special levels
"_Missing_" and "_Other_"
ylevels - mapped new levels, in the initial binning it should be same as
xlevels
good - number of good in each xlevel
bad - number of bad in each xlevel

EXAMPLE of returned object
obj <- list(xlevels = c("A", "B", "_Missing_", "_Other_"),
            ylevels = c("A", "B", "_Missing_", "_Other_"),
            good = c(200, 200, 200, 200),
            bad = c(100, 200, 300, 400))
}
\description{
Initial binning for character variables
}

