% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_binning.R
\name{bin_init_num}
\alias{bin_init_num}
\title{Initial binning for numeric variables}
\usage{
bin_init_num(sdf, x, y, init_bins = 100)
}
\arguments{
\item{sdf}{Spark data frame}

\item{x}{the name of the numeric variable to bin}

\item{y}{the name of the target variable, which should take only
values 0 and 1, no missing should be allowed}

\item{init_bins}{number of initial bins}
}
\value{
a numeric binning object with class name "intervalbin" with the
following fields:
cuts - the break points (left end should be -Inf, right end should be Inf)
good - number of good (0) in each interval
bad - number of bad (1) in each interval
missing - a list with two components for missing values: good count and
bad count

EXAMPLE returned object
obj <- list(cuts = c(-Inf, 1.5, 10, Inf),
            good = c(200, 300, 500),
            bad = c(100, 300, 900),
            _Missing_ = c(good = 50, bad = 50))
}
\description{
Initial binning for numeric variables
}
