% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_class.R
\name{bin_table.intervalbin}
\alias{bin_table.intervalbin}
\title{Generate binning table data for intervalbin class}
\usage{
\method{bin_table}{intervalbin}(x)
}
\arguments{
\item{x}{binning object}
}
\description{
Generate binning table data for intervalbin class
}

