% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_class.R
\name{bin_table.nominalbin}
\alias{bin_table.nominalbin}
\title{Generate binning table data for nominalbin class}
\usage{
\method{bin_table}{nominalbin}(x)
}
\arguments{
\item{x}{binning object}
}
\description{
Generate binning table data for nominalbin class
}

