% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_transformation.R
\name{bin_transform}
\alias{bin_transform}
\title{Transformation Function For Binning Object}
\usage{
bin_transform(sdf, binobj, input.col, output.col = paste0("ft_", input.col))
}
\arguments{
\item{sdf}{spark data frame}

\item{binobj}{character binning object with class "nominalbin"}

\item{input.col}{input feature name}

\item{output.col}{output feature name}
}
\description{
Transformation Function For Binning Object
}

