% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_transformation.R
\name{bin_transform_batch}
\alias{bin_transform_batch}
\title{Transform Function using a list of Binning Object}
\usage{
bin_transform_batch(sdf, binobjs)
}
\arguments{
\item{sdf}{spark data frame}

\item{binobjs}{binning objects}
}
\description{
Transform Function using a list of Binning Object
}
