% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_transformation.R
\name{bin_transform_char}
\alias{bin_transform_char}
\title{Optimal binning transformation for character features}
\usage{
bin_transform_char(sdf, binobj, input.col, output.col)
}
\arguments{
\item{sdf}{spark data frame}

\item{binobj}{character binning object with class "nominalbin"}

\item{input.col}{input feature name}

\item{output.col}{output feature name}
}
\value{
a new spark data frame with transformed features
}
\description{
Optimal binning transformation for character features
}

