% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_transformation.R
\name{bin_transform_num}
\alias{bin_transform_num}
\title{Optimal binning transformation for numeric features}
\usage{
bin_transform_num(sdf, binobj, input.col, output.col)
}
\arguments{
\item{sdf}{spark data frame}

\item{binobj}{numeric binning object with class "intervalbin"}

\item{input.col}{input feature name}

\item{output.col}{output feature name}
}
\value{
a new spark data frame with transformed features
}
\description{
Optimal binning transformation for numeric features
}

