% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_binning.R
\name{bin_tree.intervalbin}
\alias{bin_tree.intervalbin}
\title{Optimal binning using tree for numeric variable}
\usage{
\method{bin_tree}{intervalbin}(x, ...)
}
\arguments{
\item{x}{the initial binning object}

\item{...}{additional parameters passed to tree function call}
}
\value{
a new intervalbin object
}
\description{
Optimal binning using tree for numeric variable
}

