% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_binning.R
\name{bin_tree.nominalbin}
\alias{bin_tree.nominalbin}
\title{Optimal binning using tree for character variable}
\usage{
\method{bin_tree}{nominalbin}(x, ...)
}
\arguments{
\item{x}{the initial binning object}

\item{...}{additional parameters passed to tree function call}
}
\value{
a new nominalbin object
}
\description{
Optimal binning using tree for character variable
}
