% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_binning.R
\name{bin_tree}
\alias{bin_tree}
\title{Optimal binning using classification tree method}
\usage{
bin_tree(x, ...)
}
\arguments{
\item{x}{the initial binning object, nominalbin or intervalbin classes}

\item{...}{additional arguments}
}
\description{
Optimal binning using classification tree method
}

