% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lending}
\alias{lending}
\title{Lending Club Data}
\format{An object of class \code{data.frame} with 42538 rows and 112 columns.}
\source{
\url{https://www.lendingclub.com}
}
\usage{
lending
}
\description{
A dataset containing Lending Club data
}
\details{
\describe{
}
}
\keyword{datasets}

