% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_class.R
\name{plot.intervalbin}
\alias{plot.intervalbin}
\title{Plot intervalbin class}
\usage{
\method{plot}{intervalbin}(x)
}
\arguments{
\item{x}{intervabin class object}
}
\description{
Plot intervalbin class
}
