% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_class.R
\name{print.intervalbin}
\alias{print.intervalbin}
\title{Function to print intervalbin object}
\usage{
\method{print}{intervalbin}(x)
}
\arguments{
\item{x}{the intervalbin object}
}
\description{
Function to print intervalbin object
}

