% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_class.R
\name{print.nominalbin}
\alias{print.nominalbin}
\title{Function to print nominalbin object}
\usage{
\method{print}{nominalbin}(x)
}
\arguments{
\item{x}{the nominal object}
}
\description{
Function to print nominalbin object
}
