% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_modeling.R
\name{sparkbin_auto_model}
\alias{sparkbin_auto_model}
\title{Automatic modeling using sparkbin package}
\usage{
sparkbin_auto_model(sc, sdf, y, minIV = 0.03, ...)
}
\arguments{
\item{sc}{spark connection}

\item{sdf}{spark data frame or data frame in R}

\item{y}{the target variable with values 0/1}

\item{minIV}{the minium information value for a variable to keep}

\item{...}{additional arguments for future}
}
\description{
Automatic modeling using sparkbin package
}

